+++ librelane/steps/openroad.py
@@ -2188,6 +2188,20 @@ class WriteViews(OpenROADStep):
         return os.path.join(get_script_dir(), "openroad", "write_views.tcl")
 
 
+@Step.factory.register()
+class WriteCDL(OpenROADStep):
+    """
+    Write CDL view of an ODB design
+    """
+
+    id = "OpenROAD.WriteCDL"
+    name = "Write CDL"
+    outputs = [DesignFormat.CDL]
+
+    def get_script_path(self):
+        return os.path.join(get_script_dir(), "openroad", "write_cdl.tcl")
+
+
 # Resizer Steps